shortnews/requirements.txt
annotated-types==0.6.0
anyio==4.2.0
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
boto3==1.33.13
botocore==1.33.13
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.1
cryptography==41.0.5
cx-Oracle==8.3.0
cycler==0.12.1
Cython==0.29.21
debugpy==1.7.0
decorator==5.1.1
entrypoints==0.4
exceptiongroup==1.1.3
fastapi==0.109.2
filelock==3.12.2
fonttools==4.38.0
fsspec==2023.1.0
google-api-core==2.15.0
google-auth==2.26.1
google-cloud-texttospeech==2.15.1
googleapis-common-protos==1.62.0
grpcio==1.60.0
grpcio-status==1.60.0
h11==0.14.0
huggingface-hub==0.16.4
idna==3.6
importlib-metadata==6.7.0
ipykernel==6.16.2
ipython==7.34.0
ipywidgets==8.1.1
jedi==0.19.1
jmespath==1.0.1
joblib==1.3.2
JPype1==1.5.0
jupyter_client==7.4.9
jupyter_core==4.12.0
jupyterlab-widgets==3.0.9
keybert==0.8.3
kiwisolver==1.4.5
konlpy==0.6.0
lxml==4.9.3
markdown-it-py==2.2.0
matplotlib==3.5.3
matplotlib-inline==0.1.6
mdurl==0.1.2
nest-asyncio==1.5.8
nltk==3.8.1
numpy==1.26.4
outcome==1.3.0.post0
packaging==23.2
pandas==1.3.5
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
prompt-toolkit==3.0.41
proto-plus==1.23.0
protobuf==4.24.4
psutil==5.9.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
Pygments==2.17.2
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2023.3.post1
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.1
regex==2023.12.25
requests==2.31.0
rich==13.7.0
rsa==4.9
s3transfer==0.8.2
safetensors==0.4.1
schedule==1.2.1
scikit-learn==1.0.2
scipy==1.12.0
selenium==4.11.2
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
smart-open==6.4.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
starlette==0.36.3
style==1.1.0
summa==1.2.0
threadpoolctl==3.1.0
tokenizers==0.13.3
torch==1.13.1
torchvision==0.14.1
tornado==6.2
tqdm==4.66.1
traitlets==5.9.0
transformers==4.30.2
trio==0.22.2
trio-websocket==0.11.1
typing_extensions==4.9.0
update==0.0.1
urllib3==1.26.18
uvicorn==0.27.1
wcwidth==0.2.12
webdriver-manager==4.0.1
widgetsnbextension==4.0.9
wordcloud==1.9.3
wsproto==1.2.0
zipp==3.15.0
